기존에 만든 컴퍼넌트와 이름이 겹침
전부 하나로 수정함

+++ Unity Project/Assets/Monster/Monster_stats.cs
@@ -5,8 +5,8 @@ using UnityEditor;
 
 public class Monster_stats : MonoBehaviour
 {
-    public float Hp;
-    public float damage;
+    public int Hp;
+    public int damage;
     public float Atk_dir;
     Drop_Item item_Drop;
     private string Drop_item_it;
@@ -18,7 +18,7 @@ public class Monster_stats : MonoBehaviour
         item_Drop = new Drop_Item();
     }
 
-    void Get_damange(float damage)
+    void Get_damange(int damage)
     {
         
         Hp -= damage;
